Fix Travis CI config for Python.
On Travis CI, "python: 2.7" is not a valid Python version specification
on macOS; this is documented here:
https://docs.travis-ci.com/user/languages/python/
> Python builds are not available on the macOS environment.

and:

> 'language: python' is an error on Travis CI macOS

Using it as we do on Linux fails as follows:
https://travis-ci.org/mbrukman/csv2txf/jobs/596627885
with an error:

> Unable to download 2.7 archive. The archive may not exist. Please
> consider a different version.

Instead, we need to specify XCode versions which will select the version
of macOS we're using, which will come with a particular version of
Python.

For Linux, Python 3.4 is not supported on Ubuntu Bionic 18.04:
https://travis-ci.org/mbrukman/csv2txf/jobs/596652250
shows that the following URL returns 4o4:
https://storage.googleapis.com/travis-ci-language-archives/python/binaries/ubuntu/18.04/x86_64/python-3.4.tar.bz2
164

In fact, Travis CI says that they only support Python 3.6+ on Bionic:
https://docs.travis-ci.com/user/reference/bionic/#python-support
but Python 3.5 actually works there:
https://travis-ci.org/mbrukman/csv2txf/jobs/596651521
so we can use it, at least for now.