Add system test for `csv2txf.py`
Previously, we only had tests for parsing of transactions on a per-broker basis,
but the tool we have users using is actually `csv2txf.py` which provides the
final output that folks will import into their tax software, so we need to have
tests for that as well.

Also added regen tool to generate or update the test data, which also required
adding an injectable date parameter to the printed output to be used for
ensuring a fixed date for testing.