Added human-readable summary for CSV inputs
This change adds a flag `--outfmt` to `csv2txf.py` which defaults to the value
`txf` to maintain the current behavior of outputting the TXF formatted file
for import to TurboTax.

However, if `--outfmt=summary` is specified to `csv2txf.py` instead, the new
functionality outputs a human-readable summary to be able to easily verify the
transactions at a glance.