Update GitHub Actions to Ubuntu 22.04 and macOS 11
The previous versions were quite old, and in the case of macOS 10.15, no
longer available, leading to stuck actions are there are no available
runners to run the tests:

https://docs.github.com/en/actions/using-github-hosted-runners/about-github-hosted-runners/about-github-hosted-runners#standard-github-hosted-runners-for-public-repositories

Remove Python 3.6 as it is already EOL, and add Python 3.9.

Also add `fail-fast: false` to run all tests even if some fail, so we
can get more information about the other tests that will run in
parallel.